📁 refactor(admin): product, coupon 폴더로 디렉토리 구조 분리
