+++ src/basic/components/admin/procuct/ProductManager.tsx
